# docs/openapi.yaml
openapi: 3.1.0
info:
  title: NetVigil API
  version: 0.1.0
  description: |
    NetVigil is an AI-driven Network Detection and Response platform for SMBs.
    This document is the single source of truth for the contract between the
    FastAPI backend (`services/api`) and the TypeScript clients
    (`packages/web`, `packages/mobile`). Generated TypeScript types live in
    `packages/shared-types/src/openapi.d.ts` and must not be edited by hand.

    Authentication: bearer JWT (RS256), 15-minute access tokens with refresh
    rotation. All tenant-scoped endpoints require the bearer token.
  contact:
    name: NIT3003 Capstone Team
  license:
    name: Educational use — Victoria University
servers:
  - url: http://localhost:8000/api/v1
    description: Local dev (docker-compose)
  - url: https://staging.netvigil.example/api/v1
    description: Staging (AWS ECS)
...
paths:
...
  /users:
    get:
      tags: [admin]
      summary: List users in the caller's organisation
      parameters:
        - in: query
          name: status
          schema: { $ref: '#/components/schemas/UserStatus' }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/OrgUser' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /users/{userId}:
    parameters:
      - in: path
        name: userId
        required: true
        schema: { type: string, format: uuid }
    patch:
      tags: [admin]
      summary: Update a user's role, active flag, or status
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                role: { $ref: '#/components/schemas/UserRole' }
                isActive: { type: boolean }
                status: { $ref: '#/components/schemas/UserStatus' }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/OrgUser' }
        '404': { $ref: '#/components/responses/NotFound' }
  /users/{userId}/approve:
    parameters:
      - in: path
        name: userId
        required: true
        schema: { type: string, format: uuid }
    post:
      tags: [admin]
      summary: Approve a pending user (sets status to active)
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/OrgUser' }
        '404': { $ref: '#/components/responses/NotFound' }
  /users/{userId}/reject:
    parameters:
      - in: path
        name: userId
        required: true
        schema: { type: string, format: uuid }
    post:
      tags: [admin]
      summary: Reject a pending user
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/OrgUser' }
        '404': { $ref: '#/components/responses/NotFound' }

  # ─── Super-admin panel ────────────────────────────────────────────────────
...
  /admin/organizations/{orgId}/users:
    parameters:
      - in: path
        name: orgId
        required: true
        schema: { type: string, format: uuid }
    get:
      tags: [admin]
      summary: List all users in a specific organisation (super_admin only)
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/AdminUser' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /admin/users:
    get:
      tags: [admin]
      summary: List all users across all organisations (super_admin only)
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/AdminUser' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /admin/users/{userId}:
    parameters:
      - in: path
        name: userId
        required: true
        schema: { type: string, format: uuid }
    patch:
      tags: [admin]
      summary: Update any user across all orgs (super_admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                role: { $ref: '#/components/schemas/UserRole' }
                isActive: { type: boolean }
                status: { $ref: '#/components/schemas/UserStatus' }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AdminUser' }
        '404': { $ref: '#/components/responses/NotFound' }

  # ─── Devices (FR-02, FR-09) ──────────────────────────────────────────────
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  responses:
...
    Unauthorized:
      description: Invalid or missing credentials
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Forbidden:
      description: Insufficient permissions
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
  schemas:
    Error:
      type: object
      required: [code, message]
      properties:
        code: { type: string, example: invalid_credentials }
        message: { type: string }
        details:
          type: object
          additionalProperties: true

    # ─── Auth ────────────────────────────────────────────────────────────
...
    UserRole:
      type: string
      enum:
        - super_admin
        - admin
        - senior_analyst
        - analyst
        - threat_hunter
        - forensic_investigator
        - auditor
        - developer
    UserStatus:
      type: string
      enum: [pending, active, rejected]
...
    OrgUser:
      type: object
      required: [id, email, role, isActive, status, mfaEnrolled, createdAt]
      properties:
        id: { type: string, format: uuid }
        email: { type: string, format: email }
        role: { $ref: '#/components/schemas/UserRole' }
        isActive: { type: boolean }
        status: { $ref: '#/components/schemas/UserStatus' }
        mfaEnrolled: { type: boolean }
        createdAt: { type: string, format: date-time }
...
    AdminUser:
      type: object
      required: [id, organizationId, organizationName, email, role, isActive, status, mfaEnrolled, createdAt]
      properties:
        id: { type: string, format: uuid }
        organizationId: { type: string, format: uuid }
        organizationName: { type: string }
        email: { type: string, format: email }
        role: { $ref: '#/components/schemas/UserRole' }
        isActive: { type: boolean }
        status: { $ref: '#/components/schemas/UserStatus' }
        mfaEnrolled: { type: boolean }
        createdAt: { type: string, format: date-time }

    # ─── Devices ─────────────────────────────────────────────────────────
...
security:
  - bearerAuth: []